requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
channels==4.2.2
channels_redis==4.2.1
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2025.4
cron-descriptor==1.4.5
Django==5.2.1
django-celery==3.1.17
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-filter==25.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
kombu==5.5.3
msgpack==1.1.0
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==6.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.9.0
django-auditlog
django-jazzmin